Add docker files with additional jmeter configuration.
