+++ bootstrapping/app_aws.env
@@ -1,5 +1,5 @@
 #!/bin/bash
-export TOKEN=populatd_by_script
+export TOKEN=populated_by_script
 export AWS_ACCESS_KEY_ID=AKIAIEGP47DY3P3IKIVA
 export AWS_SECRET_ACCESS_KEY=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
 export AWS_DEFAULT_REGION=us-west-2
