+++ loadtest-library/src/main/java/org/opentestsystem/loadtesting/Content.java
@@ -1,3 +1,5 @@
+package org.opentestsystem.loadtesting;
+
 import java.util.ArrayList;
 import java.util.List;
