+++ loadtest-library/src/main/java/org/opentestsystem/loadtesting/Item.java
@@ -1,3 +1,5 @@
+package org.opentestsystem.loadtesting;
+
 /**
  * Created by emunoz on 11/27/15.
  */
