+++ loadtest-library/src/main/java/org/opentestsystem/loadtesting/PageContent.java
@@ -1,14 +1,21 @@
-import org.w3c.dom.*;
-import javax.xml.parsers.*;
+package org.opentestsystem.loadtesting;
+
+import org.w3c.dom.Document;
+import org.w3c.dom.Element;
+import org.w3c.dom.Node;
+import org.w3c.dom.NodeList;
+
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.transform.OutputKeys;
 import javax.xml.transform.Transformer;
 import javax.xml.transform.TransformerFactory;
 import javax.xml.transform.dom.DOMSource;
 import javax.xml.transform.stream.StreamResult;
-import java.io.*;
+import java.io.ByteArrayInputStream;
+import java.io.InputStream;
+import java.io.StringWriter;
 import java.nio.charset.StandardCharsets;
-import java.util.ArrayList;
-import java.util.List;
 
 public class PageContent {
     private Content _content;
@@ -127,3 +134,4 @@ public class PageContent {
 
 
 }
+
