+++ tds-jmeter-client/src/main/docker/user.properties
@@ -1,0 +1,2 @@
+httpclient4.retrycount=1
+hc.parameters.file=hc.parameters
