+++ tds-jmeter-server/src/main/docker/hc.parameters
@@ -1,0 +1,1 @@
+http.connection.stalecheck$Boolean=true
